requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
Django==4.2.10
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-mathfilters==1.0.0
djangorestframework==3.14.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
pillow==10.2.0
pycparser==2.21
PySocks==1.7.1
pytz==2024.1
reportlab==4.1.0
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
wsproto==1.2.0
